+++ teamup/src/components/home.js
@@ -1,0 +1,9 @@
+import React from "react";
+const Home = ()=>{
+
+
+    return (<>
+        <h1>홈페이지</h1>
+    </>);
+};
+export default Home;
